WIP: Continued refactoring, api and client work
