+++ client/pages/admin/users.vue
@@ -127,7 +127,7 @@ export default {
             
             // POST to /users
             try {
-                const response = await this.$axios.post('/users', axiosData, axiosConfig)
+                const response = await this.$axios.post('/users/', axiosData, axiosConfig)
 
                 // This is all for notifications of successful uploads
                 this.submittedUsers[response.data.id].submittionSuccess = true // Note success
